# src/main/java/dev/ximarelli/rosary/backend/auth/LoginRequest.java
package dev.ximarelli.rosary.backend.auth;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank @Size(min = 6, max = 100) String password) {
}